build

+++ webpack.config.js
@@ -190,22 +190,24 @@ module.exports = (env) => {
 						];
 						file = '../spontify.me-app/config.xml';
 						var s = fs.readFileSync(file, 'utf8');
-						for (var i = 0; i < regexs.length; i++) {
-							if (!regexs[i].pattern.test(s)) {
-								fs.rmdirSync('dist', { recursive: true });
-								throw new Error('regex ' + regexs[i].pattern + ' failed');
+						if (s) {
+							for (var i = 0; i < regexs.length; i++) {
+								if (!regexs[i].pattern.test(s)) {
+									fs.rmdirSync('dist', { recursive: true });
+									throw new Error('regex ' + regexs[i].pattern + ' failed');
+								}
+								s = s.replace(regexs[i].pattern, regexs[i].replace);
 							}
-							s = s.replace(regexs[i].pattern, regexs[i].replace);
+							fs.writeFileSync(file, s);
+							file = '../spontify.me-app/GoogleService-Info.plist';
+							s = fs.readFileSync(file, 'utf8');
+							s = s.replace(/(<string>com\.jq\.)([^<]+)/, '$1' + props.bundleId.substring(7));
+							fs.writeFileSync(file, s);
+							file = '../spontify.me-app/google-services.json';
+							s = fs.readFileSync(file, 'utf8');
+							s = s.replace(/("package_name": "com\.jq\.)([^"]+)/, '$1' + props.bundleId.substring(7));
+							fs.writeFileSync(file, s);
 						}
-						fs.writeFileSync(file, s);
-						file = '../spontify.me-app/GoogleService-Info.plist';
-						s = fs.readFileSync(file, 'utf8');
-						s = s.replace(/(<string>com\.jq\.)([^<]+)/, '$1' + props.bundleId.substring(7));
-						fs.writeFileSync(file, s);
-						file = '../spontify.me-app/google-services.json';
-						s = fs.readFileSync(file, 'utf8');
-						s = s.replace(/("package_name": "com\.jq\.)([^"]+)/, '$1' + props.bundleId.substring(7));
-						fs.writeFileSync(file, s);
 						file = 'dist/images/logo.svg';
 						s = fs.readFileSync(file, 'utf8')
 							.replace('{placeholderAppTitle}', props.nameShort);
